+++ examples/src/store.js
@@ -1,0 +1,8 @@
+import { createQueryStore } from './hooks'
+
+export default createQueryStore({
+  auth: {
+    token: null,
+    user: null,
+  }
+});