+++ src/index.tsx
@@ -1,45 +1,62 @@
-import { useState, useEffect, useRef, createContext, useContext, useCallback } from 'react';
+// @ts-ignore
+import { useState, useEffect, useRef, createContext, useContext, useCallback, Dispatch, SetStateAction, MutableRefObject } from 'react';
 
-/**
- * @callback QueryResultCacheFn
- * @param {string} url
- * 
- * @typedef {object} QueryResultCache
- * @property {QueryResultCacheFn} get
- * @property {QueryResultCacheFn} has
- * @property {QueryResultCacheFn} clear
- *
- * @typedef {object} QueryResult
- * @property {*} data
- * @property {Error|false} error
- * @property {boolean} loading
- * @property {function} updateData
- * @property {function} refresh
- * @property {QueryResultCache} cache
- *
- * @typedef {object} QueryEventResult
- * @property {function} sendRequest
- * @property {boolean} isSending
- * @property {*} data
- * @property {Error|false} error
- * @property {boolean} loading
- * @property {function} updateData
- * @property {function} refresh
- * @property {QueryResultCache} cache
- * 
- * @typedef {object} DefaultUseQueryOptions
- * @property {string} [name] 
- * @property {string} [selector] 
- * @property {array|function} [pick]
- * @property {function} [transform] 
- * @property {string} [method="GET"] 
- * @property {object} [headers] 
- * @property {*} [body] 
- * @property {boolean} [isDebuggerActivated=false] 
- * @property {number} [cacheTimeout=0] 
- */
+interface QueryResultCacheFn {
+    get: Function,
+    has: Function,
+    clear: Function,
+}
 
-const DEFAULT_OPTIONS = {
+interface QueryResult {
+    data: any;
+    error: Error | false | null;
+    loading: boolean;
+    mutate: Function;
+    refresh: Function;
+    cache: QueryResultCacheFn;
+}
+
+interface QueryEventResult {
+    sendRequest: Function;
+    isSending: boolean;
+    data: any;
+    error: Error | false | null;
+    loading: boolean;
+    mutate: Function;
+    refresh: Function;
+    cache: QueryResultCacheFn;
+}
+
+interface DefaultUseQueryOptions {
+    name?: string;
+    selector?: string;
+    pick?: Array<any> | Function;
+    transform?: Function;
+    method?: string;
+    headers?: object;
+    body?: any;
+    isDebuggerActivated?: boolean;
+    cacheTimeout?: number;
+}
+
+interface CacheObject {
+    current: object
+}
+
+interface CreateStoreReturn {
+    __value: object
+}
+
+type DispatchersActions = {
+    actions: object
+}
+
+type DispatcherOptions = {
+    name: string,
+    actions: object
+}
+
+const DEFAULT_OPTIONS: DefaultUseQueryOptions = {
     name: undefined,
     selector: undefined,
     transform: undefined,
@@ -51,7 +68,15 @@ const DEFAULT_OPTIONS = {
     cacheTimeout: 0,
 };
 
-let cacheTimeout = null;
+const dispatchers: DispatchersActions = {
+    actions: {}
+};
+
+const cache: CacheObject = {
+    current: {}
+}
+
+let cacheTimeout: number | null = null;
 
 /**
  * @param {string} [message=""] 
@@ -59,7 +84,7 @@ let cacheTimeout = null;
  * @param {boolean} [isError=false] 
  * @param {null|string|Error} [error=null]
  */
-const logDebugger = (message = "", isDebuggerActivated = false, isError = false, error = null) => {
+const logDebugger = (message: string = "", isDebuggerActivated: boolean = false, isError: boolean = false, error: null | Error = null) => {
     if (isError) {
         if (isDebuggerActivated) console.error(`[USE_QUERY]: ${message} - ${new Date().toLocaleTimeString()}`, error);
     } else {
@@ -71,22 +96,22 @@ const logDebugger = (message = "", isDebuggerActivated = false, isError = false,
  * @param {DefaultUseQueryOptions} options 
  * @returns {boolean}
  */
-const validateOptions = (options) => {
+const validateOptions = (options: DefaultUseQueryOptions): boolean => {
     const keys = Object.keys(DEFAULT_OPTIONS);
     const _keys = Object.keys(options);
     const symDifference = _keys.filter(item => !keys.includes(item));
 
     if (options.selector !== undefined && typeof options.selector !== "string") {
-        logDebugger("Validation error for `options.selector`", isDebuggerActivated, true, new Error("Selector must be a string. It should contains key value to select from result object."));
+        logDebugger("Validation error for `options.selector`", options.isDebuggerActivated, true, new Error("Selector must be a string. It should contains key value to select from result object."));
         return false;
     } else if (options.name !== undefined && typeof options.name !== "string") {
-        logDebugger("Validation error for `options.name`", isDebuggerActivated, true, new Error("Name must be a string. It should contains data name to retrieve it from state."));
+        logDebugger("Validation error for `options.name`", options.isDebuggerActivated, true, new Error("Name must be a string. It should contains data name to retrieve it from state."));
         return false;
     } else if (options.method !== undefined && (typeof options.method !== "string" || ["GET", "POST", "PUT", "PATCH", "DELETE"].indexOf(options.method.toUpperCase()) === -1)) {
-        logDebugger("Validation error for `options.method`", isDebuggerActivated, true, new Error('Method must be a string. It should be one of "GET", "POST", "PUT", "PATCH", "DELETE".'));
+        logDebugger("Validation error for `options.method`", options.isDebuggerActivated, true, new Error('Method must be a string. It should be one of "GET", "POST", "PUT", "PATCH", "DELETE".'));
         return false;
     } else if (options.headers !== undefined && typeof options.headers !== "object" && Array.isArray(options.headers)) {
-        logDebugger("Validation error for `options.headers`", isDebuggerActivated, true, new Error("Headers must be an object. It should contains request headers key value."));
+        logDebugger("Validation error for `options.headers`", options.isDebuggerActivated, true, new Error("Headers must be an object. It should contains request headers key value."));
         return false;
     } else if (options.isDebuggerActivated !== undefined && typeof options.isDebuggerActivated !== "boolean" && Array.isArray(options.headers)) {
         logDebugger("Validation error for `options.isDebuggerActivated`", true, true, new Error("isDebuggerActivated must be a boolean. It should be activated if you need to debug all process."));
@@ -114,36 +139,136 @@ const validateOptions = (options) => {
  * @param {DefaultUseQueryOptions} options 
  * @param {MutableRefObject} cache 
  */
-const timeoutCacheClear = (options, cache) => {
+const timeoutCacheClear = (options: DefaultUseQueryOptions, cache: MutableRefObject<any>) => {
     if (options.cacheTimeout && options.cacheTimeout > 0) {
         logDebugger(`Cache clear timeout start: ${options.cacheTimeout / 1000} seconds.`, options.isDebuggerActivated);
-        clearTimeout(cacheTimeout);
-        cacheTimeout = setTimeout(() => {
+        window.clearTimeout(cacheTimeout as number);
+        cacheTimeout = window.setTimeout(() => {
             cache.current = {};
             logDebugger(`Cache cleared`, options.isDebuggerActivated);
         }, options.cacheTimeout);
     }
 }
 
+/**
+ * Creates a store with the provided initial state.
+ * @param {object} initialState - The initial state object for the store.
+ * @returns {CreateStoreReturn} An object representing the created store.
+ */
+export const createQueryStore = (initialState: object): CreateStoreReturn => {
+    Object.entries(initialState).forEach(([key, value]) => {
+        cache.current[key] = value;
+    });
+
+    return { __value: initialState };
+}
+
 const QueryContext = createContext(null);
 
-export const QueryProvider = ({ children }) => {
-    const [state, setState] = useState(null);
+export const QueryProvider = ({ children, store, dispatchers: _dispatchers }) => {
+    if (store && !store.__value) throw new Error("`store` must be a valid QueryStore data, use `createQueryStore` function to create it");
+
+    if (_dispatchers && !Array.isArray(_dispatchers)) {
+        throw new Error("`dispatchers` must be a valid array of dispatchers, use `createDispatcher` function to generate them");
+    } else {
+        _dispatchers.forEach(({ name, actions }) => {
+            dispatchers.actions = {
+                ...dispatchers.actions,
+                [name]: {
+                    ...dispatchers.actions[name],
+                    ...actions,
+                },
+            }
+        });
+    }
+    
+    const [state, setState] = useState(store ? store.__value : null);
+
     return (
+        // @ts-ignore
         <QueryContext.Provider value={[state, setState]}>
             {children}
         </QueryContext.Provider>
     )
 }
 
-export const useQueryContext = (name = undefined) => {
+/**
+ * Creates a dispatcher object with specified actions for managing state updates.
+ * @param {DispatcherOptions} dispatcherOptions - Options for configuring the dispatcher.
+ * @returns {DispatcherObject} An object representing the created dispatcher.
+ */
+export const createQueryDispatcher = (dispatcherOptions: DispatcherOptions) => {
+    const actions = {};
+
+    Object.entries(dispatcherOptions.actions).forEach(([actionName, actionFn]) => {
+        actions[actionName] = (data: any, setData: Function) => (payload: any) => {
+            const _data = { ...data }[dispatcherOptions.name];
+            
+            actionFn(_data, payload);
+
+            setData({ ...data, ..._data });
+        };
+    });
+
+    return {
+        name: dispatcherOptions.name,
+        actions
+    }
+}
+
+/**
+ * Custom hook for accessing dispatcher actions associated with a specific name.
+ * @param {string} name - The name of the dispatcher to retrieve actions for.
+ * @returns {Object} An object containing the actions associated with the specified dispatcher name.
+ */
+export const useQueryDispatcher = (name: string) => {
+    // @ts-ignore
+    const [data, setData] = useContext(QueryContext);
+
+    const actions = {};
+    
+    Object.entries(dispatchers.actions[name]).forEach(([actionName, actionFn]) => {
+        // @ts-ignore
+        actions[actionName] = (payload: any) => actionFn(data, setData)(payload);
+    });
+
+    return actions;
+}
+
+/**
+ * Custom hook for accessing and updating query state from the QueryContext.
+ * @param {string} [name=undefined] - The name of the specific state to access within the query data.
+ * @returns {[S, Dispatch<SetStateAction<S>>]} An array containing the queried state and a function to update the state.
+ *                 If `name` is provided, it returns the specified state and a function to update that state.
+ *                 If `name` is not provided, it returns the entire query data and a function to update it.
+ */
+export const useQueryState = (name: string | undefined = undefined): [any, Dispatch<SetStateAction<any>>] => {
+    // @ts-ignore
     const [data, setData] = useContext(QueryContext);
     
     if (name) { 
-        return [data ? data[name] : data, (value) => setData((prev) => ({ ...prev, [name]: value }))];
+        return [
+            data ? data[name] : data, 
+            (value) => {
+                setData((prev: any) => ({ ...prev, [name]: value }));
+                cache.current[name] = value;
+            }
+        ];
     } else {
         return [data, setData];
     }
+};
+
+/**
+ * Custom hook for querying data from the QueryContext and applying a callback selector.
+ * @param {function} [callbackSelector] - The callback function to apply as a selector on the data.
+ * @returns {*} The result of applying the callback selector on the data from the QueryContext.
+ */
+export const useQuerySelector = (callbackSelector: Function = (state: any) => {}): any => {
+    // @ts-ignore
+    const [data] = useContext(QueryContext);
+    
+    return callbackSelector(data);
 };
 
 /**
@@ -161,14 +286,14 @@ export const useQueryContext = (name = undefined) => {
  * @param {number} [options.cacheTimeout] cacheTimeout must be a number. It rappresent the timeout to remove cached data from memory in milliseconds.
  * @returns {QueryResult}
  */
-export const useQuery = (url, options = { ...DEFAULT_OPTIONS }) => {
+export const useQuery = (url: string, options: DefaultUseQueryOptions = { ...DEFAULT_OPTIONS }): QueryResult => {
     const { name, selector, pick, transform, method, headers, body, isDebuggerActivated } = { ...DEFAULT_OPTIONS, ...options };
 
     validateOptions({ ...DEFAULT_OPTIONS, ...options });
     
-    const cache = useRef({});
+    // const cache = useRef({});
 
-    const [data, setData] = useQueryContext(name);
+    const [data, setData] = useQueryState(name);
     const [error, setError] = useState(null);
     const [loading, setLoading] = useState(false);
 
@@ -192,7 +317,8 @@ export const useQuery = (url, options = { ...DEFAULT_OPTIONS }) => {
         try {
             logDebugger("No data found in cache, proceed to do a new request...", isDebuggerActivated);
 
-            const response = await fetch(url, { method, headers, data: body });
+            // @ts-ignore
+            const response = await fetch(url, { method, headers, body });
             const result = await response.json();
 
             if (response.ok) {
@@ -201,6 +327,7 @@ export const useQuery = (url, options = { ...DEFAULT_OPTIONS }) => {
                 let _transformedData = typeof transform === "function" ? transform(_result) : _result;
                 if (typeof pick === "function" || Array.isArray(pick)) {
                     logDebugger("pick data before saving", isDebuggerActivated);
+                    // @ts-ignore
                     _transformedData = JSON.parse(JSON.stringify(_transformedData, pick));
                 }
                 setData(_transformedData);
@@ -261,16 +388,16 @@ export const useQuery = (url, options = { ...DEFAULT_OPTIONS }) => {
  * @param {number} [options.cacheTimeout] cacheTimeout must be a number. It rappresent the timeout to remove cached data from memory in milliseconds.
  * @returns {QueryEventResult}
  */
-export const useQueryEvent = (url, options = { ...DEFAULT_OPTIONS }) => {
+export const useQueryEvent = (url: string, options: DefaultUseQueryOptions = { ...DEFAULT_OPTIONS }): QueryEventResult => {
     const [isSending, setIsSending] = useState(false);
 
     const { name, selector, pick, transform, method, headers, body, isDebuggerActivated } = { ...DEFAULT_OPTIONS, ...options };
 
     validateOptions({ ...DEFAULT_OPTIONS, ...options });
     
-    const cache = useRef({});
+    // const cache = useRef({});
 
-    const [data, setData] = useQueryContext(name);
+    const [data, setData] = useQueryState(name);
     const [error, setError] = useState(null);
     const [loading, setLoading] = useState(false);
 
@@ -294,6 +421,7 @@ export const useQueryEvent = (url, options = { ...DEFAULT_OPTIONS }) => {
         try {
             logDebugger("No data found in cache, proceed to do a new request...", isDebuggerActivated);
 
+            // @ts-ignore
             const response = await fetch(url, { method, headers, data: body });
             const result = await response.json();
 
@@ -303,6 +431,7 @@ export const useQueryEvent = (url, options = { ...DEFAULT_OPTIONS }) => {
                 let _transformedData = typeof transform === "function" ? transform(_result) : _result;
                 if (typeof pick === "function" || Array.isArray(pick)) {
                     logDebugger("pick data before saving", isDebuggerActivated);
+                    // @ts-ignore
                     _transformedData = JSON.parse(JSON.stringify(_transformedData, pick));
                 }
                 setData(_transformedData);
@@ -343,10 +472,10 @@ export const useQueryEvent = (url, options = { ...DEFAULT_OPTIONS }) => {
         mutate: setData,
         refresh: fetchData,
         cache: {
-            get: (url) => {
+            get: (url: string) => {
                 return url ? cache.current[url] : cache.current;
             },
-            has: (url) => {
+            has: (url: string) => {
                 return cache.current[url];
             },
             clear: () => {
@@ -355,4 +484,4 @@ export const useQueryEvent = (url, options = { ...DEFAULT_OPTIONS }) => {
             }
         }
     }
-  }
+}