+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [2.0.0] 2024-03-13
+
+Introduces global state managment with `createDispatcher` and `useDispatcher`.
+
+## [1.3.0] 2024-03-13
+
+Renamed `useQueryContext` to `useQueryState`, added `useQuerySelector` hook and rewrite all in typescript.
+
 ## [1.2.0] 2024-03-13
 
 Added `useQueryEvent` hook to trigger request on browser event like button `click`.